Fix int64 Spectral Embedding (#7689)
We need to edit the coo wrapper api (rows, cols, vals) to use int64 for the indexes and also the python nnz value.

Authors:
  - Anupam (https://github.com/aamijar)

Approvers:
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/7689

+++ cpp/src/spectral/spectral_embedding.cu
@@ -45,9 +45,9 @@ void transform(raft::resources const& handle,
 
 void transform(raft::resources const& handle,
                ML::SpectralEmbedding::params config,
-               raft::device_vector_view<int, int> rows,
-               raft::device_vector_view<int, int> cols,
-               raft::device_vector_view<float, int> vals,
+               raft::device_vector_view<int, int64_t> rows,
+               raft::device_vector_view<int, int64_t> cols,
+               raft::device_vector_view<float, int64_t> vals,
                raft::device_matrix_view<float, int, raft::col_major> embedding)
 {
   auto connectivity_graph_view = raft::make_device_coo_matrix_view<float, int, int, int64_t>(
